requirements.txt
bcrypt==4.1.2
blinker==1.7.0
click==8.1.7
configparser==6.0.0
exceptiongroup==1.2.0
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mysql-connector==2.2.9
mysqlclient==2.2.3
packaging==23.2
pluggy==1.4.0
PyJWT==2.8.0
PyMySQL==1.1.0
pytest==8.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.25
tomli==2.0.1
typing_extensions==4.9.0
Werkzeug==3.0.1
